+++ test/iterable.test.ts
@@ -1,36 +1,9 @@
 import { ReadableStream } from "node:stream/web";
 import { vi } from "vitest";
 import { iterable } from "../src";
-import { serializeIterableValue } from "../src/utils/internal/iterable";
 import { describeMatrix } from "./_setup";
 
 describeMatrix("iterable", (t, { it, expect, describe }) => {
-  // TODO: move to unit
-  describe("serializeIterableValue", () => {
-    const exampleDate: Date = new Date(Date.UTC(2015, 6, 21, 3, 24, 54, 888));
-    it.each([
-      { value: "Hello, world!", output: "Hello, world!" },
-      { value: 123, output: "123" },
-      { value: 1n, output: "1" },
-      { value: true, output: "true" },
-      { value: false, output: "false" },
-      { value: undefined, output: undefined },
-      { value: null, output: "null" },
-      { value: exampleDate, output: JSON.stringify(exampleDate) },
-      { value: { field: 1 }, output: '{"field":1}' },
-      { value: [1, 2, 3], output: "[1,2,3]" },
-      { value: () => {}, output: undefined },
-      {
-        value: Buffer.from("Hello, world!"),
-        output: Buffer.from("Hello, world!"),
-      },
-      { value: Uint8Array.from([1, 2, 3]), output: Uint8Array.from([1, 2, 3]) },
-    ])("$value => $output", ({ value, output }) => {
-      const serialized = serializeIterableValue(value);
-      expect(serialized).toStrictEqual(output);
-    });
-  });
-
   // TODO: Investigate issue with iterable oon web (Received non-Uint8Array chunk)
   describe.skipIf(t.target === "web")("iterable", () => {
     it("sends empty body for an empty iterator", async () => {
